third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations-expected.txt

Tests the Timeline API instrumentation of style recalc events with invalidations.

PASS
PASS
PASS


Running: testLocalFrame
first recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:9}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:10}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: multipleStyleRecalcs
first recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:20}
        changedAttribute : undefined
        changedClass : "green"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='green'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:20}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='green'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
second recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:22}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:22}
        changedAttribute : undefined
        changedClass : "green"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:23}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:23}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
third recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:25}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:25}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:26}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:26}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:27}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementThree' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testSubframe
first recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:38}
        changedAttribute : undefined
        changedClass : "green"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='green'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
second recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:40}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:40}
        changedAttribute : undefined
        changedClass : "green"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:41}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
third recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:43}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:43}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:44}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:44}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations.html:45}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementThree' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
